# db/schemas/swagger.yaml
swagger: "2.0"
info:
  description: "These APIs cater to UTV platform"
  version: "1.0.0"
  title: "User To Views"
  contact:
    email: "guptaaditya24@gmail.com"
host: "localhost:3000"
basePath: "/"
...
schemes:
- "https"
- "http"
paths:
...
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "This can only be done by the logged out user."
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User details object"
        required: true
        schema:
          $ref: "#/definitions/User-signup"
      responses:
        200:
          description: "successfully created user"
        400:
          description: "Invalid username/password/name"
        403:
          description: "User with the username is already registered"
        default:
          description: "successfully created user"
    put:
      tags:
      - "user"
      summary: "Update user password"
      description: "This can only be done by the logged in user, i.e. with the Bearer token"
      operationId: "updateUser"
      parameters:
      - in: "header"
        name: "Authorization"
        type: "string"
        format: "Bearer"
      - in: "body"
        name: "body"
        description: "User password"
        schema:
          $ref: "#/definitions/User-update"
      responses:
        200:
          description: "Updated user details"
        401:
          description: "Invalid token/Token expired"
        403:
          description: "Invalid token. Not authorized for the operation"
    get:
      tags:
      - "user"
      summary: "Get requesting user details"
      description: "This can be done only by logged in user"
      operationId: "getUserDetails"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "Authorization"
        type: "string"
        format: "Bearer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User-details"
        401:
          description: "User unauthorized"
        400:
          description: "Invalid token"
...
  /link:
    get:
      tags:
      - "links"
      summary: "Get list of links created by user"
      description: "Retrieve all the links created by user"
      operationId: "getAllLinks"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "Authorization"
        type: "string"
        format: "Bearer"
      responses:
        200:
          description: "User created links"
          schema:
            $ref: "#/definitions/LinksList"
        401:
          description: "User unauthorized"
        400:
          description: "Invalid token"
    post:
      tags:
      - "links"
      summary: "Create shortened link with overlay and targetUrl"
      description: "This can only be called by logged in user"
      operationId: "createNewLink"
      parameters:
      - in: "header"
        name: "Authorization"
        type: "string"
        format: "Bearer"
      - in: "body"
        name: "body"
        description: "Require targetUrl, overlay (optional) to create a new link"
        required: true
        schema:
          $ref: "#/definitions/CreateLink"
      responses:
        200:
          description: "Created link details"
          schema:
            $ref: "#/definitions/Link"
        400:
          description: "Invalid token"
...
definitions:
...
  User-details:
    type: "object"
    properties:
      username:
        type: "string"
      name:
        type: "string"
  User-signup:
    type: "object"
    properties:
      username:
        type: "string"
      firstname:
        type: "string"
      lastname:
        type: "string"
      password:
        type: "string"
...
  User-update:
    type: "object"
    properties:
      password:
        type: "string"
      name:
        type: "string"
      timeone:
        type: "string"
    required: 
    - name
...
  LinksList:
    type: "array"
    items:
      $ref: "#/definitions/Link"
...
  Link:
    type: "object"
    properties:
      id:
        type: "integer"
      shortUrl:
        type: "string"
        example: "http://utv.com/v/pksd5ds"
      visits:
        type: "integer"
      overlayId:
        type: "string"
      targetUrl:
        type: "string"
      createdOn:
        type: "string"
        format: "date-time"
    required:
      - id
      - shortUrl
      - targetUrl
      - createdOn
  CreateLink:
    type: "object"
    properties:
      overlayId:
        type: "string"
      targetUrl:
        type: "string"
    required:
      - targetUrl
...